fix binary test
