+++ pkg/config/validation/validation.go
@@ -30,7 +30,6 @@ import (
 	udpaa "github.com/cncf/xds/go/udpa/annotations"
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 	"github.com/gogo/protobuf/types"
-	"github.com/google/cel-go/cel"
 	"github.com/hashicorp/go-multierror"
 	"github.com/lestrrat-go/jwx/jwt"
 	"google.golang.org/protobuf/proto"
@@ -3435,17 +3434,6 @@ func validateTelemetryAccessLogging(logging []*telemetry.AccessLogging) (v Valid
 		v = appendValidation(v, validateTelemetryProviders(l.Providers))
 	}
 	return
-}
-
-func validateTelemetryFilter(filter *telemetry.AccessLogging_Filter) error {
-	expr := filter.Expression
-	env, _ := cel.NewEnv()
-	_, issue := env.Parse(expr)
-	if issue.Err() != nil {
-		return fmt.Errorf("must be a valid CEL expression, %w", issue.Err())
-	}
-
-	return nil
 }
 
 func validateTelemetryTracing(tracing []*telemetry.Tracing) (v Validation) {
